Fix imports after moving recent changes in SFCC connector to SDK
